Share the webhook between v1 and v2 API versions
